implement downvoting on answers on a question show page

+++ app/controllers/votes_controller.rb
@@ -29,7 +29,7 @@ end
 delete '/questions/:voteable_id/votes' do
   question = Question.find(params[:voteable_id])
   votes = Vote.where(voteable_id: params[:voteable_id], voteable_type: "Question")
-  if votes.length > 0
+  if votes.length > 0 && session[:user_id]
     votes.last.destroy
   end
   question.points.to_json
@@ -43,6 +43,16 @@ post '/questions/:question_id/answers/:answer_id/votes' do
                 voteable_id: params[:answer_id], 
                 voteable_type: "Answer",
                 user_id: session[:user_id])
+  end
+  answer.votes.length.to_json
+end
+
+delete '/questions/:question_id/answers/:answer_id/votes' do
+  question = Question.find(params[:question_id])
+  answer = Answer.find(params[:answer_id])
+  votes = Vote.where(voteable_id: params[:answer_id], voteable_type: "Answer")
+  if votes.length > 0 && session[:user_id]
+    votes.last.destroy
   end
   answer.votes.length.to_json
 end
